requirements.txt
streamlit
pandas
browser-use
python-dotenv
